move files to their own places.
