+++ app/src/main/java/com/nima/tmdb/business/data/network/implementation/RemoteDataSourceImp.kt
@@ -12,7 +12,7 @@ import com.nima.tmdb.business.domain.model.requests.WatchlistBody
 import com.nima.tmdb.business.domain.model.responses.FavoriteResponse
 import com.nima.tmdb.business.domain.model.trend.TrendInfoModel
 import com.nima.tmdb.framewrok.datasource.network.abstraction.RemoteService
-import com.nima.tmdb.requests.wrapper.ApiWrapper
+import com.nima.tmdb.business.data.network.requests.wrapper.ApiWrapper
 
 class RemoteDataSourceImp(
     private val remoteService: RemoteService
