+++ app/src/main/java/com/nima/tmdb/business/domain/model/login/account/Account.kt
@@ -79,8 +79,8 @@ class Account {
     var statusCode: Int? = null
 }
 
-fun Account.asDatabaseAccount()  : com.nima.tmdb.database.entities.Account {
-    return com.nima.tmdb.database.entities.Account(
+fun Account.asDatabaseAccount()  : com.nima.tmdb.framewrok.datasource.cache.entities.Account {
+    return com.nima.tmdb.framewrok.datasource.cache.entities.Account(
         id = this.id!!,
         hash = this.avatar?.getGravatar()?.getHash(),
         iso6391 = this.iso6391,
