+++ app/src/main/java/com/nima/tmdb/business/domain/util/TimeWindow.kt
@@ -1,0 +1,6 @@
+package com.nima.tmdb.business.domain.util
+
+sealed class TimeWindow {
+    data class Day(private val time : String = "day"): TimeWindow()
+    data class Week(private val time : String = "week"): TimeWindow()
+}
