--- app/src/main/java/com/nima/tmdb/business/domain/util/helper/TimeWindow.kt
@@ -1,6 +1,0 @@
-package com.nima.tmdb.business.domain.util.helper
-
-sealed class TimeWindow {
-    data class Day(private val time : String = "day"):TimeWindow()
-    data class Week(private val time : String = "week"):TimeWindow()
-}
