+++ app/src/main/java/com/nima/tmdb/business/interactors/createdList/GetCreatedList.kt
@@ -3,7 +3,7 @@ package com.nima.tmdb.business.interactors.createdList
 import com.nima.tmdb.business.data.cache.abstraction.CacheDataSource
 import com.nima.tmdb.business.data.network.abstraction.RemoteDataSource
 import com.nima.tmdb.business.domain.model.account.lists.CreatedLists
-import com.nima.tmdb.requests.wrapper.ApiWrapper
+import com.nima.tmdb.business.data.network.requests.wrapper.ApiWrapper
 
 class GetCreatedList(
     val cache: CacheDataSource,
