+++ app/src/main/java/com/nima/tmdb/business/interactors/details/GetMovieById.kt
@@ -3,7 +3,7 @@ package com.nima.tmdb.business.interactors.details
 import com.nima.tmdb.business.data.cache.abstraction.CacheDataSource
 import com.nima.tmdb.business.data.network.abstraction.RemoteDataSource
 import com.nima.tmdb.business.domain.model.Details
-import com.nima.tmdb.requests.wrapper.ApiWrapper
+import com.nima.tmdb.business.data.network.requests.wrapper.ApiWrapper
 
 class GetMovieById(
     private val cache: CacheDataSource,
