+++ app/src/main/java/com/nima/tmdb/business/interactors/login/Login.kt
@@ -4,7 +4,7 @@ import com.nima.tmdb.business.data.cache.abstraction.CacheDataSource
 import com.nima.tmdb.business.data.network.abstraction.RemoteDataSource
 import com.nima.tmdb.business.domain.model.login.LoginInfo
 import com.nima.tmdb.business.domain.model.login.LoginResponse
-import com.nima.tmdb.requests.wrapper.ApiWrapper
+import com.nima.tmdb.business.data.network.requests.wrapper.ApiWrapper
 
 class Login(
     private val cache: CacheDataSource,
