+++ app/src/main/java/com/nima/tmdb/business/interactors/mainPage/MarkAsFavorite.kt
@@ -4,7 +4,7 @@ import com.nima.tmdb.business.data.cache.abstraction.CacheDataSource
 import com.nima.tmdb.business.data.network.abstraction.RemoteDataSource
 import com.nima.tmdb.business.domain.model.requests.FavoriteBody
 import com.nima.tmdb.business.domain.model.responses.FavoriteResponse
-import com.nima.tmdb.requests.wrapper.ApiWrapper
+import com.nima.tmdb.business.data.network.requests.wrapper.ApiWrapper
 
 class MarkAsFavorite(
     private val cache: CacheDataSource,
