+++ app/src/main/java/com/nima/tmdb/framewrok/datasource/cache/database/MyDao.kt
@@ -1,12 +1,12 @@
-package com.nima.tmdb.database
+package com.nima.tmdb.framewrok.datasource.cache.database
 
 import androidx.lifecycle.LiveData
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.nima.tmdb.database.entities.Account
-import com.nima.tmdb.database.entities.UserInfo
+import com.nima.tmdb.framewrok.datasource.cache.entities.Account
+import com.nima.tmdb.framewrok.datasource.cache.entities.UserInfo
 
 @Dao
 interface MyDao {
