+++ app/src/main/java/com/nima/tmdb/framewrok/datasource/cache/database/TMDBDataBase.kt
@@ -1,9 +1,9 @@
-package com.nima.tmdb.database
+package com.nima.tmdb.framewrok.datasource.cache.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.nima.tmdb.database.entities.Account
-import com.nima.tmdb.database.entities.UserInfo
+import com.nima.tmdb.framewrok.datasource.cache.entities.Account
+import com.nima.tmdb.framewrok.datasource.cache.entities.UserInfo
 
 @Database(
     entities = [
