+++ app/src/main/java/com/nima/tmdb/framewrok/presentation/createdlist/CreatedListViewModel.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.nima.tmdb.business.domain.model.account.lists.CreatedLists
 import com.nima.tmdb.business.interactors.createdList.CreatedListInteractors
-import com.nima.tmdb.requests.wrapper.ApiWrapper
+import com.nima.tmdb.business.data.network.requests.wrapper.ApiWrapper
 import kotlinx.coroutines.launch
 
 class CreatedListViewModel @ViewModelInject constructor(
