+++ app/src/main/java/com/nima/tmdb/framewrok/presentation/login/LoginFragment.kt
@@ -15,13 +15,13 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
 import com.nima.tmdb.R
-import com.nima.tmdb.database.MyDao
+import com.nima.tmdb.framewrok.datasource.cache.database.MyDao
 import com.nima.tmdb.databinding.FragmentLoginBinding
 import com.nima.tmdb.business.domain.model.login.LoginInfo
 import com.nima.tmdb.business.domain.model.login.LoginResponse
 import com.nima.tmdb.business.domain.model.login.RequestToken
 import com.nima.tmdb.business.domain.model.login.Session
-import com.nima.tmdb.requests.wrapper.ApiWrapper
+import com.nima.tmdb.business.data.network.requests.wrapper.ApiWrapper
 import com.nima.tmdb.utils.Constants.API_KEY
 import com.nima.tmdb.utils.toast
 import dagger.hilt.android.AndroidEntryPoint
