+++ app/src/main/java/com/nima/tmdb/framewrok/presentation/splash/SplashFragment.kt
@@ -20,7 +20,7 @@ import com.nima.tmdb.business.domain.model.login.LoginResponse
 import com.nima.tmdb.business.domain.model.login.RequestToken
 import com.nima.tmdb.business.domain.model.login.Session
 import com.nima.tmdb.databinding.FragmentSplashBinding
-import com.nima.tmdb.requests.wrapper.ApiWrapper
+import com.nima.tmdb.business.data.network.requests.wrapper.ApiWrapper
 import com.nima.tmdb.utils.Constants.API_KEY
 import com.nima.tmdb.utils.toast
 import dagger.hilt.android.AndroidEntryPoint
